chore: review tweaks
