+++ src/client/src/App/Trades/__tests__/FxTrades.core.test.tsx
@@ -9,10 +9,9 @@ import {
 import { BehaviorSubject, Subject } from "rxjs"
 
 import { Trade, tradesTestData } from "@/services/trades"
-import { TestThemeProvider } from "@/utils/testUtils"
+import { setupMockWindow, TestThemeProvider } from "@/utils/testUtils"
 
 import FxTrades from "../CoreFxTrades"
-import { setupWindow } from "./utils"
 
 jest.mock("@/services/trades/trades")
 jest.mock("../TradesState/tableTrades", () => ({
@@ -33,7 +32,7 @@ const renderComponent = () =>
 const _trades = require("@/services/trades/trades")
 
 describe("Trades", () => {
-  setupWindow()
+  setupMockWindow()
 
   beforeEach(() => {
     _trades.__resetMocks()
